+++ CotEditor/Sources/SidebarViewController.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2016-2018 1024jp
+//  © 2016-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -76,7 +76,7 @@ final class SidebarViewController: NSTabViewController {
     }
     
     
-    /// restore last state
+    /// keys to be restored from the last session
     override class var restorableStateKeyPaths: [String] {
         
         return super.restorableStateKeyPaths + [